handle no-account scenario more explicitly
